Move functions to a dedicated file
