+++ src/seed_tools/commands/split_seed_json.ts
@@ -5,10 +5,7 @@
 
 import { Command } from '@commander-js/extra-typings';
 import { promises as fs } from 'fs';
-import DefaultMap from '../../base/containers/default_map';
-import { type Study } from '../../proto/generated/study';
-import { VariationsSeed } from '../../proto/generated/variations_seed';
-import * as study_json_utils from '../utils/study_json_utils';
+import { parseLegacySeedJson } from '../utils/legacy_json_to_seed';
 
 export default function createCommand() {
   return new Command('split_seed_json')
@@ -16,26 +13,6 @@ export default function createCommand() {
     .argument('<seed_json_path>', 'path to seed.json')
     .argument('<output_dir>', 'output directory')
     .action(main);
-}
-
-export function parseLegacySeedJson(seedContent: string): {
-  parsedSeed: VariationsSeed;
-  studiesMap: DefaultMap<string, Study[]>;
-} {
-  const seedJson = preprocessSeedJson(JSON.parse(seedContent));
-
-  // Parse the seed as protobuf json representation. The parse will fail if any
-  // unknown fields or values are present in the json.
-  const parsedSeed = VariationsSeed.fromJson(seedJson, {
-    ignoreUnknownFields: false,
-  });
-
-  const studiesMap = new DefaultMap<string, Study[]>(() => []);
-  for (const study of parsedSeed.study) {
-    studiesMap.get(study.name).push(study);
-  }
-
-  return { parsedSeed, studiesMap };
 }
 
 async function main(seedPath: string, outputDir: string) {
@@ -58,45 +35,4 @@ async function main(seedPath: string, outputDir: string) {
     const studyFile = `${outputDir}/${studyName}.json5`;
     await study_json_utils.writeStudyFile(studyArray, studyFile);
   }
-}
-
-function preprocessSeedJson(json: any): any {
-  json.study = json.studies;
-  delete json.studies;
-
-  for (const study of json.study) {
-    if (study.experiments !== undefined) {
-      study.experiment = study.experiments;
-      delete study.experiments;
-    }
-    for (const experiment of study.experiment) {
-      if (experiment.parameters !== undefined) {
-        experiment.param = experiment.parameters;
-        delete experiment.parameters;
-      }
-    }
-    if (study.filter !== undefined) {
-      if (study.filter.channel !== undefined) {
-        study.filter.channel = study.filter.channel.map((channel: string) => {
-          switch (channel) {
-            case 'NIGHTLY':
-              return 'CANARY';
-            case 'RELEASE':
-              return 'STABLE';
-            default:
-              return channel;
-          }
-        });
-      }
-      if (study.filter.platform !== undefined) {
-        study.filter.platform = study.filter.platform.map(
-          (platform: string) => {
-            return 'PLATFORM_' + platform;
-          },
-        );
-      }
-    }
-  }
-
-  return json;
 }
